+++ Sources/Sentry/SentrySdkInfo.m
@@ -1,40 +1,17 @@
 #import "SentrySdkInfo.h"
 #import "SentryClient+Private.h"
+#import "SentryExtraPackages.h"
 #import "SentryHub+Private.h"
 #import "SentryMeta.h"
 #import "SentryOptions.h"
 #import "SentrySDK+Private.h"
+#import "SentrySdkPackage.h"
 #import "SentrySwift.h"
 #import <Foundation/Foundation.h>
-
-typedef NS_ENUM(NSUInteger, SentryPackageManagerOption) {
-    SentrySwiftPackageManager,
-    SentryCocoaPods,
-    SentryCarthage,
-    SentryPackageManagerUnkown
-};
-
-/**
- * This is required to identify the package manager used when installing sentry.
- */
-#if SWIFT_PACKAGE
-static SentryPackageManagerOption SENTRY_PACKAGE_INFO = SentrySwiftPackageManager;
-#elif COCOAPODS
-static SentryPackageManagerOption SENTRY_PACKAGE_INFO = SentryCocoaPods;
-#elif CARTHAGE_YES
-// CARTHAGE is a xcodebuild build setting with value `YES`, we need to convert it into a compiler
-// definition to be able to use it.
-static SentryPackageManagerOption SENTRY_PACKAGE_INFO = SentryCarthage;
-#else
-static SentryPackageManagerOption SENTRY_PACKAGE_INFO = SentryPackageManagerUnkown;
-#endif
 
 NS_ASSUME_NONNULL_BEGIN
 
 @interface SentrySdkInfo ()
-
-@property (nonatomic) SentryPackageManagerOption packageManager;
-
 @end
 
 @implementation SentrySdkInfo
@@ -59,23 +36,32 @@ NS_ASSUME_NONNULL_BEGIN
     }
 #endif
 
+    NSMutableSet<NSDictionary<NSString *, NSString *> *> *packages =
+        [SentryExtraPackages getPackages];
+    NSDictionary<NSString *, NSString *> *sdkPackage = [SentrySdkPackage global];
+    if (sdkPackage != nil) {
+        [packages addObject:sdkPackage];
+    }
+
     return [self initWithName:SentryMeta.sdkName
                       version:SentryMeta.versionString
                  integrations:integrations
-                     features:features];
+                     features:features
+                     packages:[packages allObjects]];
 }
 
 - (instancetype)initWithName:(NSString *)name
                      version:(NSString *)version
                 integrations:(NSArray<NSString *> *)integrations
                     features:(NSArray<NSString *> *)features
+                    packages:(NSArray<NSDictionary<NSString *, NSString *> *> *)packages
 {
     if (self = [super init]) {
         _name = name ?: @"";
         _version = version ?: @"";
-        _packageManager = SENTRY_PACKAGE_INFO;
         _integrations = integrations ?: @[];
         _features = features ?: @[];
+        _packages = packages ?: @[];
     }
 
     return self;
@@ -87,6 +73,7 @@ NS_ASSUME_NONNULL_BEGIN
     NSString *version = @"";
     NSMutableSet<NSString *> *integrations = [[NSMutableSet alloc] init];
     NSMutableSet<NSString *> *features = [[NSMutableSet alloc] init];
+    NSMutableSet<NSDictionary<NSString *, NSString *> *> *packages = [[NSMutableSet alloc] init];
 
     if ([dict[@"name"] isKindOfClass:[NSString class]]) {
         name = dict[@"name"];
@@ -112,48 +99,32 @@ NS_ASSUME_NONNULL_BEGIN
         }
     }
 
+    if ([dict[@"packages"] isKindOfClass:[NSArray class]]) {
+        for (id item in dict[@"packages"]) {
+            if ([item isKindOfClass:[NSDictionary class]] &&
+                [item[@"name"] isKindOfClass:[NSString class]] &&
+                [item[@"version"] isKindOfClass:[NSString class]]) {
+                [packages addObject:@{ @"name" : item[@"name"], @"version" : item[@"version"] }];
+            }
+        }
+    }
+
     return [self initWithName:name
                       version:version
                  integrations:[integrations allObjects]
-                     features:[features allObjects]];
-}
-
-- (nullable NSString *)getPackageName:(SentryPackageManagerOption)packageManager
-{
-    switch (packageManager) {
-    case SentrySwiftPackageManager:
-        return @"spm:getsentry/%@";
-    case SentryCocoaPods:
-        return @"cocoapods:getsentry/%@";
-    case SentryCarthage:
-        return @"carthage:getsentry/%@";
-    default:
-        return nil;
-    }
+                     features:[features allObjects]
+                     packages:[packages allObjects]];
 }
 
 - (NSDictionary<NSString *, id> *)serialize
 {
-    NSMutableDictionary *sdk = @{
+    return @{
         @"name" : self.name,
         @"version" : self.version,
         @"integrations" : self.integrations,
         @"features" : self.features,
-    }
-                                   .mutableCopy;
-    if (self.packageManager != SentryPackageManagerUnkown) {
-        NSString *format = [self getPackageName:self.packageManager];
-        if (format != nil) {
-            sdk[@"packages"] = @[
-                @{
-                    @"name" : [NSString stringWithFormat:format, self.name],
-                    @"version" : self.version
-                },
-            ];
-        }
-    }
-
-    return sdk;
+        @"packages" : self.packages,
+    };
 }
 
 @end